requirements.txt
# Django & API
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0

# Configuration
python-decouple>=3.8,<4.0

# Database (PostgreSQL driver)
psycopg[binary]>=3.1,<4.0

# Production server (optional local run in Docker)
gunicorn>=21.0,<23.0

# Testing
pytest>=7.4,<9.0
pytest-django>=4.5,<5.0
